+++ plugin-explorer/src/app/app.component.ts
@@ -15,7 +15,6 @@ import { PluginFilter, PluginService } from './plugin.service';
 import { Plugin } from './plugin-info';
 import { checked, d, setChecked } from './utilities/dom';
 import { MessageType, sendMessage } from './utilities/messages';
-import { TestFilter, getTestFilters } from './test-filter';
 
 // In order to use the Webview UI Toolkit web components they
 // must be registered with the browser (i.e. webview) using the
@@ -42,9 +41,10 @@ export class AppComponent implements OnInit {
   terms = '';
   listTitle = '';
   isInstalled: string | undefined = 'true';
-  testFilters: TestFilter[] = getTestFilters();
+  isCapOnly: string | undefined = 'false';
 
   count = 0;
+  assetsUri = '';
   busy = true;
   private searchedPlugin: Plugin | undefined; // Found from search
 
@@ -59,6 +59,7 @@ export class AppComponent implements OnInit {
     switch (event.data.command) {
       case MessageType.getPlugins:
         await this.pluginService.get(event.data.uri);
+        this.assetsUri = event.data.assetsUri;
         this.pluginService.calculatedUnknownPlugins();
         this.search();
         this.busy = false;
@@ -117,9 +118,8 @@ export class AppComponent implements OnInit {
       this.listTitle = 'Official Plugins';
     }
 
-    const checkedTestsTitle = this.checkedTestsTitle();
-    if (checkedTestsTitle != '') {
-      this.listTitle = `Plugins that work with ${checkedTestsTitle}`;
+    if (checked('capOnly')) {
+      this.listTitle = `Plugins excluding Cordova plugins`;
     }
 
     const android = checked('android');
@@ -127,7 +127,7 @@ export class AppComponent implements OnInit {
     const both = checked('both');
     const any = checked('any');
 
-    this.plugins = this.pluginService.search(filters, this.terms, this.checkedTests(), android, ios, both, any);
+    this.plugins = this.pluginService.search(filters, this.terms, checked('capOnly'), android, ios, both, any);
     this.count = this.plugins.length;
 
     if (filters.includes(PluginFilter.search)) this.listTitle += ` related to '${this.terms}'`;
@@ -144,25 +144,5 @@ export class AppComponent implements OnInit {
       console.log(`Send request for ${this.terms}`);
       sendMessage(MessageType.getPlugin, this.terms);
     }
-  }
-
-  private checkedTestsTitle(): string {
-    let result: string[] = [];
-    for (const testFilter of this.testFilters) {
-      if (checked(testFilter.id)) {
-        result = [...result, testFilter.name];
-      }
-    }
-    return result.join(' or ');
-  }
-
-  private checkedTests(): string[] {
-    let result: string[] = [];
-    for (const testFilter of this.testFilters) {
-      if (checked(testFilter.id)) {
-        result = [...result, ...testFilter.list];
-      }
-    }
-    return result;
   }
 }