+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - You can now press ALT+B to build your project
 - If computer has multiple IP Addresses then remember the IP Address chosen
 - Fix to check for the version of Java
+- Fix issue with detection of missing plugin for Typescript wrappers
 
 ### Version 1.40